% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spearmantest.R
\name{spearmantest}
\alias{spearmantest}
\title{Calculates Spearman's correlation}
\usage{
spearmantest(X, Y)
}
\arguments{
\item{X}{A vector of numbers}

\item{Y}{A vector of numbers}
}
\value{
The spearman correlation between \code{X} and \code{Y}
}
\description{
Calculates Spearman's correlation
}
\examples{
spearmantest(c(1,2,3), c(6,5,4))
spearmantest(10, 1)

}
