% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wilcoxon.R
\name{wilcoxon}
\alias{wilcoxon}
\title{Wilcoxon Statistic and Test}
\usage{
wilcoxon(X, Y, alpha = 0.05, alternative = "two.sided", digits = 4)
}
\arguments{
\item{X}{A vector of numbers}

\item{Y}{A vector of numbers}

\item{alpha}{Alpha value}

\item{alternative}{One of 'two.sided' (default), 'greater', or 'less'}

\item{digits}{The desired number of digits after the decimal point}
}
\value{
The wilcoxon test statistic and p-value between \code{X} and \code{Y}, calculated both exactly and asymptotically
}
\description{
The Wilcoxon test statistic and corresponding p-value, from one of three alternative hypotheses
}
\examples{
wilcoxon(c(1,2,3), c(6,5,4))
wilcoxon(rnorm(), rnorm())

}
