% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spearmantest.R
\name{spearmantest}
\alias{spearmantest}
\title{Calculates Spearman's correlation, which is the same as calculating Pearson correlation on the ranked data.}
\usage{
spearmantest(X, Y)
}
\arguments{
\item{X}{A vector of numbers}

\item{Y}{A vector of numbers}
}
\value{
The spearman correlation between \code{X} and \code{Y}
}
\description{
Calculates Spearman's correlation, which is the same as calculating Pearson correlation on the ranked data.
}
\examples{
add(1, 1)
add(10, 1)

}
